Create Random
This program uses a Random object to generate a random number between 1 and 10. Then, it prompts the user to guess the number using a Scanner object to read user input.
Inside a try-catch block, it attempts to read an integer input from the user. 
If the user enters a non-numeric value, an InputMismatchException is caught and handled, prompting the user to enter a numeric value.